Handle inventory views with slot/size desync
The player inventory has a size of 41, made up from four
rows of pure inventory and five slots in armor and offhand content.
When the server opens a player inventory, it uses a generic 9x4 chest
menu, which only represents 36 of the 41 items, leaving out armor and
offhand.

No network slots are registered for these slots either, meaning that
slot 37 is already a slot in the bottom inventory, as with any other
normal inventory that has a dedicated menu type. The server however uses
Inventory#getSize to split the incoming network slot ids, leading to the
server treating slot 37 as part of the top player inventory, as it is
less than 41.

This patch forces inventory views to also define the network slot
count in their top inventory, which is for the most part just their
size. The CraftContainer implementation however, being the only possible
one wrapping a CraftInventoryPlayer instance, special cases the player
inventory and now correctly returns 36, allowing a correct split of
incoming raw network slot ids.